requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.8.0
google-auth>=2.29.0
google-api-python-client>=2.125.0
python-dotenv>=1.0.1
